Basic Table Operations Input Tsv Parser TreeTableView Adduct View Next Button RT in Minutes Fixed EIC Open multiple mzXML Color! Width! Default and set color, width Comments Simple Progress Bar / Threading Started on Batch Tab Batch Tabs Renaming Simple peak picking xAxis correction (all adducts have same min max) More peak picking Correlation more peak picking... ... no more peak picking, comparison of AVGEIC generation of Interpolated EICs Quality of peak in the middle Visualization of ideal peak Method restructuring Backup before Datastructure rework rework... Matrix for shortest path Optimized Path calculation Path visualization, Peak Prop Vis Toggle shiftview Fixes, load mzxml context menu, focus on clicked adduct in adductview (needs fixes) Before Dataset-Maps After Dataset-Maps Acitve Dataset fixed currentdataset not being initialized changed/fixed POM. netbeans has problems with maven+javafx, should work now. Follow-up problems not foreseeable. Wavelet peak picking M/Z chart shows ppm M/Z shift calculation/cleaning per file Fixed ChartGenerator Loops Score for selected Files Delete Files Delete Batches minimized RAM fixed Parsing fixed File Deletion Deactivate Files Change Properties for selected files Highlight selected files in Adductview, select new ones Change color, width of files changes adductview/shift. Unselect in shiftview Fixed Adductview, no selection, selection update Hover, shiftview Open Adduct from shiftview Fixed Listener memomryleak scroll to selected OGroup when double click in Shiftview, default sort by RT Accordion rework... ...done, draw multiple Datasets, select across Datasets included RT-range for shift calculation, commented out Active Dataset De/activate whole Dataset Parameters additional scores Naive peak preparation Naive peak picking and fitting, output to console Output to file, additional scores in Table Proper sorting and start on Adductview labels fixed additional scores, min sorting for Scorepeakclose select Opacity Mode in Shiftview Experimental Distance Opacity Mode No need for PropArray Change PeakPicking and Baseline on the fly Smooth PropArray Fixed missing M/Z values/chart Selection Rectangle in Shiftview Set Penalty Area Progress bar for Adductview Disable Charts in Adductview Show peak borders Fix Output, fix Adductview, fix peak borders fixed issue with RT Tolerance>2 small changes Visualize Penalty Area Started with Peak matrix Visualization Moved peak vis, small fixes Implemented certainty small fixes, fleshed out certainty show avg. certainty and %peaks found in Batch show files in peak matrix vis started on peak Mode peak Mode implemented, only visual peak Mode manual peak, only visual peak Mode manual peak, peak picking addPeak Button fixed manual peak picking included cutoff in manual peak picking, only look for peak in inner 70% range changed peak marker, don't delete manual peaks during automatic peak picking, don't add manual peak if peak is already there delete peaks small fix more adducts delete empty slices, double->float

+++ src/main/java/com/mycompany/fxmltableview/datamodel/Dataset.java
@@ -9,10 +9,10 @@ import com.mycompany.fxmltableview.gui.BatchController;
 import com.mycompany.fxmltableview.logic.Session;
 import java.io.File;
 import java.util.List;
-import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.FloatProperty;
 import javafx.beans.property.Property;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleFloatProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
@@ -37,10 +37,10 @@ public class Dataset {
     
     //Default values for new Files
     private Property<Color> color;
-    private DoubleProperty Width;
+    private FloatProperty Width;
     
     //penalty for path calculation
-    private SimpleDoubleProperty penalty;
+    private SimpleFloatProperty penalty;
     
     private Property<Boolean> active;
     private BatchController controller;
@@ -53,9 +53,9 @@ public class Dataset {
         
         this.listofFiles = FXCollections.observableArrayList();
         //default values for new Dataset
-        this.Width = new SimpleDoubleProperty(1.5);
+        this.Width = new SimpleFloatProperty(1.5f);
         this.color = new SimpleObjectProperty(Color.BLACK);
-        this.penalty = new SimpleDoubleProperty(0.5);
+        this.penalty = new SimpleFloatProperty(0.5f);
         active = new SimpleBooleanProperty(true);
         
     }
@@ -112,37 +112,37 @@ public class Dataset {
     /**
      * @return the width
      */
-    public double getWidth() {
+    public float getWidth() {
         return Width.get();
     }
 
     /**
      * @param width the width to set
      */
-    public void setWidth(DoubleProperty width) {
+    public void setWidth(FloatProperty width) {
         this.Width = width;
     }
     
-    public DoubleProperty getWidthProperty() {
+    public FloatProperty getWidthProperty() {
         return Width;
     }
 
-    public double getPenalty() {
+    public float getPenalty() {
         return this.penalty.get();
     }
     
     /**
      * @return the penalty
      */
-    public SimpleDoubleProperty getPenaltyProperty() {
+    public SimpleFloatProperty getPenaltyProperty() {
         return penalty;
     }
 
     /**
      * @param penalty the penalty to set
      */
-    public void setPenalty(double penalty) {
-        this.penalty = new SimpleDoubleProperty(penalty);
+    public void setPenalty(float penalty) {
+        this.penalty = new SimpleFloatProperty(penalty);
     }
     
     public final Boolean getActive() {
