+++ src/main/java/com/mycompany/fxmltableview/datamodel/Peak.java
@@ -12,36 +12,36 @@ package com.mycompany.fxmltableview.datamodel;
  * hold information of a peak detected by MassSpecWavelet
  */
 public class Peak {
-    private int index;
-    private int start;
-    private int end;
-    private double scale;
-    private double SNR;
-    private double area;
+    private short index;
+    private short start;
+    private short end;
+    private float scale;
+    private float SNR;
+    private float area;
     private Slice slice;
     private boolean manual;
     
     
-    public Peak(int index, double scale, double SNR, double area, Slice slice) {
+    public Peak(short index, float scale, float SNR, float area, Slice slice) {
         this.index = index;
         this.scale = scale;
         this.SNR = SNR;
         this.area = area;
         this.slice = slice;
         this.manual = false;
-        this.start = (int) (index-1.5*scale);
+        this.start =(short) (index-1.5*scale);
         if (start<0) {
             start = 0;
         }
-        this.end = (int) (index+1.5*scale);
+        this.end = (short) (int) (index+1.5*scale);
         if (end >=slice.getFile().getSession().getResolution()) {
-            end = slice.getFile().getSession().getResolution()-1;
+            end = (short) (slice.getFile().getSession().getResolution()-1);
         }
         calculateArea();
         
     }
 
-    public Peak(int index, int start, int end, Slice slice) {
+    public Peak(short index, short start, short end, Slice slice) {
         this.index = index;
         this.start = start;
         this.end = end;
@@ -52,7 +52,7 @@ public class Peak {
         
     }
     
-    public Peak(boolean manual, int index, int start, int end, Slice slice, int non) {
+    public Peak(boolean manual, short index, short start, short end, Slice slice, int non) {
         this.index = index;
         this.start = start;
         this.end = end;
@@ -67,56 +67,56 @@ public class Peak {
     /**
      * @return the index
      */
-    public int getIndex() {
+    public short getIndex() {
         return index;
     }
 
     /**
      * @param index the index to set
      */
-    public void setIndex(int index) {
+    public void setIndex(short index) {
         this.index = index;
     }
 
     /**
      * @return the scale
      */
-    public double getScale() {
+    public float getScale() {
         return scale;
     }
 
     /**
      * @param scale the scale to set
      */
-    public void setScale(double scale) {
+    public void setScale(float scale) {
         this.scale = scale;
     }
 
     /**
      * @return the SNR
      */
-    public double getSNR() {
+    public float getSNR() {
         return SNR;
     }
 
     /**
      * @param SNR the SNR to set
      */
-    public void setSNR(double SNR) {
+    public void setSNR(float SNR) {
         this.SNR = SNR;
     }
 
     /**
      * @return the area
      */
-    public double getArea() {
+    public float getArea() {
         return area;
     }
 
     /**
      * @param area the area to set
      */
-    public void setArea(double area) {
+    public void setArea(float area) {
         this.area = area;
     }
 
@@ -137,10 +137,10 @@ public class Peak {
     public void trimPeak() {
         //max distance from middle to end in minutes
         //TODO: as parameter
-        int maxdist = slice.getFile().getSession().getMaxPeakLengthint();
-        double[] intensity = slice.getIntensityArray();
+        short maxdist = slice.getFile().getSession().getMaxPeakLengthint();
+        int[] intensity = slice.getIntensityArray();
         if (index-getStart()>maxdist) {
-            setStart(index-maxdist);
+            setStart((short) (index-maxdist));
             
             //look for minima in EIC
             while (intensity[start+1]<intensity[start]&&start<index) {
@@ -149,7 +149,7 @@ public class Peak {
             
         }
         if (getEnd()-index>maxdist) {
-            setEnd(index+maxdist);
+            setEnd((short) (index+maxdist));
             
             //look for minima in EIC
             while (intensity[end-1]<intensity[end]&&end>index) {
@@ -184,7 +184,7 @@ public class Peak {
     /**
      * @param start the start to set
      */
-    public void setStart(int start) {
+    public void setStart(short start) {
         this.start = start;
     }
 
@@ -198,7 +198,7 @@ public class Peak {
     /**
      * @param end the end to set
      */
-    public void setEnd(int end) {
+    public void setEnd(short end) {
         this.end = end;
     }
